change dto

+++ src/main/java/com/example/msinternauticacrm/controllers/BoatController.java
@@ -31,19 +31,19 @@ public class BoatController {
 
 
     @GetMapping("/get-by-id/{id}")
-    public ResponseEntity<Boat> getBoatById(@PathVariable Long id) {
+    public ResponseEntity<Boat> getBoatById(@PathVariable int id) {
         Boat boat = boatService.getBoatById(id);
         return new ResponseEntity<>(boat, HttpStatus.OK);
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<Boat> updateBoat(@PathVariable Long id, @Valid @RequestBody Boat boatDetails) {
+    public ResponseEntity<Boat> updateBoat(@PathVariable int id, @Valid @RequestBody Boat boatDetails) {
         Boat updatedBoat = boatService.updateBoat(id, boatDetails);
         return new ResponseEntity<>(updatedBoat, HttpStatus.OK);
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<String> deleteBoat(@PathVariable Long id) {
+    public ResponseEntity<String> deleteBoat(@PathVariable int id) {
         boolean isDeleted = boatService.deleteBoat(id);
         if (isDeleted) {
             return new ResponseEntity<>("Boat deleted successfully.", HttpStatus.OK);
