+++ src/main/java/com/example/msinternauticacrm/repositories/BoatRepository.java
@@ -11,10 +11,10 @@ import java.util.List;
 @org.springframework.stereotype.Repository
 public interface BoatRepository extends Repository<Boat, String>, MongoRepository<Boat, String>, PagingAndSortingRepository<Boat, String> {
 
-    Boat findBoatsByIdBoat(Long idBoat);
-    boolean existsBoatByIdBoat(Long idBoat);
+    Boat findBoatsByIdBoat(int idBoat);
+    boolean existsBoatByIdBoat(int idBoat);
 
-    void deleteBoatByIdBoat(Long idBoat);
+    void deleteBoatByIdBoat(int idBoat);
 
 
 }
